update: Send 32 bits ecg data

+++ app/source/main.c
@@ -151,7 +151,7 @@ int main(void)
   application_timers_start();
 //  advertising_start();
 
-  int16_t signal_val[ADS_NUM_CHANNEL];
+  int32_t signal_val[ADS_NUM_CHANNEL];
   int16_t ecg_channel_buf[200];
   int16_t index = 0;
